feat: #4 글올리기 모달을 BeforeModal, AfterModal component로 나눈 뒤, UploadModal component에서 합침
